refactor(Board.java): 말 갯수를 세는 메서드 반환값 수정
말 갯수를 세는 메서드 반환값을 정수로 수정

BREAKING CHANGE: 메서드 반환값을 double에서 int로 수정
as-is:
public class Board {
    ...

    public double countPiece(final PieceType pieceType, final Color color) {
        return (double) pieces.values()
                .stream()
                .filter(piece -> piece.isSamePiece(pieceType) && piece.isSameColor(color))
                .count();
    }

    ...
}

to-be:
public class Board {
    ...

    public int countPiece(final PieceType pieceType, final Color color) {
        return (int) pieces.values()
                .stream()
                .filter(piece -> piece.isSamePiece(pieceType) && piece.isSameColor(color))
                .count();
    }

    ...
}

+++ src/main/java/chess/domain/piece/PieceType.java
@@ -17,7 +17,7 @@ public enum PieceType {
         this.score = score;
     }
 
-    public double calculateScore(final double count) {
+    public double calculateScore(final int count) {
         return score * count;
     }
